mark next trial as `sent` in the model logic, not view
First, it does not belong to views logically,
Second, we will be changing the participant's stage for the trials that are last in their block. We will only do this for trials that are about to be sent. Adding this to the view logic would certainly not make any sense. So, we have to move setting `sent` out of views as well.

Because there are occasions when I need to get a new trial for internal use (mostly in development, of course) `get_next_trial` gets a new flag `about_to_be_sent` which is False by default but is set to True in the view that uses `get_next_trial` before sending its results to the client.