fix: make `n_blocks` nullable
When we create trials, we need the participant to be already in the database for which we need `n_blocks` which we won't have until we create trials.
We could save the participant inside `create_trials` but this is a very non-obvious place to do so.
So, we will just allow n_blocks to be null.
The sequence is then: create participant, create trials updating `n_blocks` once done.