fix: save all results
`trial.results.trial_id` was updated only during `trail.setup` which is run when page is loaded or if the participant exits fullscreen.
Because of this, the following trials were sent with the trial id of the first trial loaded after page load or exiting fullscreen.
Now `trial.results.trial_id` is updated during `trail.run` so that it corresponds to the last run trial.