finish blocks when we get to the new one
Before, we would finish blocks when sending out the last trial in block.
This approach broke with the very last trial, because we finished the last block but we didn't know that there were no more trials because we did not look for another one.
So, we continued on to an empty (n + 1)-th out of n block which immediately terminated (because there were no more trials) and continued to the goodbye page.
Now, we finish a block when we get to a trial from the next one.
This way, we never finish the last block and just go to the next stage.