fix: prepopulate participant form in test mode
The `participant_form` would not use the data from the instance if the request was a POST request which is standard.
The problem was that we use POST request for telling the server that the participant saw a given page.
So, `participant_form` would see a POST request and try to populate the form based on the data sent as it should when the POST comes from the form.
To fix this, we added a hidden field "form_name" to the form and check for it in the `participant_form` view.
This way, `participant_form` will ignore any POST request not coming from the form page.