added a layout for all non-experiment pages
All the content is in a single center-justified bootstrap column whose width changes with the dimensions of the screen (5/12 for large ones, 8/12 for medium, 10/12 for small, 12/12 for very  small).

After the content, there is a button that sends a post request with a single "just_saw" field to the server informing it that the participant has just seen a page corresponding to some stage and does not need to see it again.