use a list of sound urls from the server
Not the one hardcoded in the 'HeadphoneCheck.js'. It is not any less hardcoded in the view, of course. The point of this is to then change the urls to point to the files on our server instead of the McDermott's lab's S3 bucket. The problem with the bucket files is that one of the files is corrupted.