fix: avoid showing the same trial twice
This happened if `trial.save_trial_results` did not get new setting nor did it redirect.
This shouldn't normally happen and only revealed itself when I made a mistake when adding a before-redirect modal.
Nevertheless, it is better to be prepared for some possible mistakes I don't see right now.
If they happen, the screen will stay empty which will likely make the user reload the page which is ok.